refactor: move getFlashcards code out of index.js
